pyproject.toml
[project]
name = "factor-engine"
version = "0.1.0"
description = "Chilean multi-factor equity research engine"
authors = [{name = "Systematic Investing Team", email = "quant@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.6",
    "typer>=0.12",
    "SQLAlchemy>=2.0",
    "pandas>=2.1",
    "numpy>=1.26",
    "yfinance>=0.2",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "rich>=13.7",
    "requests>=2.31",
    "tabulate>=0.9",
    "notion-client>=2.2",
    "streamlit>=1.30",
    "plotly>=5.18"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.10",
    "ruff>=0.2",
    "mypy>=1.8"
]

[project.scripts]
"factor-engine" = "factor_engine.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP", "B", "W"]
ignore = ["E203"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
testpaths = ["tests"]
